refactor: split CLI into command classes
