+++ src/adapters/node-mapper.ts
@@ -129,10 +129,12 @@ const mappers = {
         createdAt: new Date(record.createdAt),
         updatedAt: new Date(record.updatedAt),
         isPublic: record.isPublic,
-        data: {
-          name: record.tagNode.name,
-          description: record.tagNode.description,
-        },
+          data: {
+            name: record.tagNode.name,
+            ...(record.tagNode.description !== null && {
+              description: record.tagNode.description,
+            }),
+          },
       }),
     toTypeRecord: (node: TagNode): Omit<TagNodeRecord, 'nodeId'> => ({
       name: node.data.name,
